Done the migrations of 'Alunos' e 'Professores'
