+++ src/main/java/com/hbsis/boletim/ligaçãoAlunoTurma/AlunoTurmaDTO.java
@@ -1,0 +1,4 @@
+package com.hbsis.boletim.ligaçãoAlunoTurma;
+
+public class AlunoTurmaDTO {
+}
