+++ src/main/java/com/hbsis/boletim/ligaçãoAlunoTurma/AlunoTurmaService.java
@@ -1,0 +1,4 @@
+package com.hbsis.boletim.ligaçãoAlunoTurma;
+
+public class AlunoTurmaService {
+}
