--- src/main/java/com/hbsis/boletim/sala/SalaRepository.java
@@ -1,6 +1,0 @@
-package com.hbsis.boletim.sala;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface SalaRepository extends JpaRepository<Sala, Long>{
-}
