+++ src/main/java/com/hbsis/boletim/turma/TurmaDTO.java
@@ -2,15 +2,15 @@ package com.hbsis.boletim.turma;
 
 public class TurmaDTO {
     private long id;
-    private String alunos;
+    private long escola;
     private String numeroTurma;
     private String periodo;
 
     public TurmaDTO(){}
 
-    public TurmaDTO(long id, String alunos, String numeroTurma, String periodo) {
+    public TurmaDTO(long id, long escola, String numeroTurma, String periodo) {
         this.id = id;
-        this.alunos = alunos;
+        this.escola = escola;
         this.numeroTurma = numeroTurma;
         this.periodo = periodo;
     }
@@ -18,7 +18,7 @@ public class TurmaDTO {
     public static TurmaDTO of(Turma turma) {
         return new TurmaDTO(
                 turma.getId(),
-                turma.getAlunos(),
+                turma.getEscola().getId(),
                 turma.getNumeroTurma(),
                 turma.getPeriodo()
         );
@@ -30,14 +30,6 @@ public class TurmaDTO {
 
     public void setId(long id) {
         this.id = id;
-    }
-
-    public String getAlunos() {
-        return alunos;
-    }
-
-    public void setAlunos(String alunos) {
-        this.alunos = alunos;
     }
 
     public String getNumeroTurma() {
@@ -56,11 +48,19 @@ public class TurmaDTO {
         this.periodo = periodo;
     }
 
+    public long getEscola() {
+        return escola;
+    }
+
+    public void setEscola(long escola) {
+        this.escola = escola;
+    }
+
     @Override
     public String toString() {
         return "TurmaDTO{" +
                 "id=" + id +
-                ", alunos='" + alunos + '\'' +
+                ", escola=" + escola +
                 ", numeroTurma='" + numeroTurma + '\'' +
                 ", periodo='" + periodo + '\'' +
                 '}';
