--- src/main/resources/db/migration/V20200210143147__salas.sql
@@ -1,10 +1,0 @@
-CREATE TABLE salas(
-
-    id BIGINT IDENTITY(1,1) NOT NULL,
-    nome_sala VARCHAR (100) NOT NULL,
-    codigo_sala VARCHAR (100) NOT NULL PRIMARY KEY,
-    escola BIGINT NOT NULL
-)
-
-ALTER TABLE salas
-ADD CONSTRAINT "For1" FOREIGN KEY ("escola") REFERENCES "escolas" (id) ON DELETE CASCADE;
