+++ src/main/resources/db/migration/V20200211091840__alunos.sql
@@ -1,0 +1,7 @@
+CREATE TABLE alunos(
+    id BIGINT IDENTITY(1,1) NOT NULL,
+    nome_aluno VARCHAR (100) NOT NULL,
+    responsavel VARCHAR (100),
+    telefone VARCHAR (13)
+)
+
