+++ src/main/resources/db/migration/V20200211091845__professores.sql
@@ -1,0 +1,8 @@
+CREATE TABLE professores(
+
+    id BIGINT IDENTITY(1,1),
+    codigo_professor VARCHAR (10),
+    nome_professor VARCHAR (100),
+    disciplina VARCHAR (25),
+    telefone VARCHAR (13)
+);
